SoIT-chatbot/requirements.txt
langchain-community
langchain
sentence-transformers
qdrant-client
python-dotenv
google-generativeai
PyPDF2
langchain-qdrant